refactor: split StatsScreen and MileageScreen god files, dedup chart cards and efficiency formula
- StatsScreen.kt: extract RecordsCard + pager machinery to stats/RecordsCard.kt
  and the three dialogs (SyncLogsDialog, GapRecordDialog, RangeRecordDialog)
  to stats/StatsDialogs.kt
- MileageScreen.kt: one file per overlay level (YearDetailScreen.kt,
  MonthDetailScreen.kt, DayDetailScreen.kt) plus MileageSharedComponents.kt
  with SummaryRow/StatChip and a single parameterized MileageChartCard
  replacing the three near-identical Yearly/Monthly/DailyChartCard copies
  (shared y-axis formatter defined once)
- Extract the repeated (totalEnergy * 1000.0) / totalDistance efficiency
  formula (7 call sites across MileageScreen files and MileageViewModel)
  into internal efficiencyWhPerUnit(), preserving the kWh->Wh-not-a-unit-
  conversion documentation
- Code moved verbatim; visibility widened private->internal only where
  symbols are now referenced cross-file. Zero behavior change.
